Fix B3 sampling API to support missing values
Motivation:

B3 sampling handling was supposed to support both present and not values.
However, in practise it only supported present values, and treated non existing as sampling=off.

Modifications:

Support Boolean values for handling sampling which now allows for not present values in the carrier,
and the local sampler can make the final decision on whether to enable sampling for the local span or not.
The change unfortunately broke API because of the type change, thus we took advantage of that and cleaned up
some problematic relationships between classes which results in a cleaner more consistent API, which also
follows closer the spec.

Result:

Correct sampling & cleaner API